+++ tests/testthat/test-ggwithinstats.R
@@ -1,5 +1,3 @@
-context(desc = "ggwithinstats")
-
 # for t-test
 data_bugs_2 <- ggstatsplot::bugs_long %>%
   dplyr::filter(.data = ., condition %in% c("HDLF", "HDHF"))
@@ -13,40 +11,42 @@ testthat::test_that(
 
     # plot
     set.seed(123)
-    p1 <- ggstatsplot::ggwithinstats(
-      data = data_bugs_2,
-      x = condition,
-      y = desire,
-      type = "bayes",
-      sort = "descending",
-      sort.fun = mean,
-      k = 4,
-      conf.level = 0.99,
-      outlier.tagging = TRUE,
-      outlier.label = "region",
-      outlier.coef = 1.5,
-      bf.message = TRUE,
-      pairwise.comparisons = TRUE,
-      pairwise.annotation = "asterisk",
-      title = "bugs dataset",
-      caption = "From `jmv` package",
-      messages = FALSE
-    )
+    p1 <-
+      ggstatsplot::ggwithinstats(
+        data = data_bugs_2,
+        x = condition,
+        y = desire,
+        type = "bayes",
+        sort = "descending",
+        sort.fun = mean,
+        k = 4,
+        conf.level = 0.99,
+        outlier.tagging = TRUE,
+        outlier.label = "region",
+        outlier.coef = 1.5,
+        bf.message = TRUE,
+        pairwise.comparisons = TRUE,
+        pairwise.annotation = "asterisk",
+        title = "bugs dataset",
+        caption = "From `jmv` package",
+        messages = FALSE
+      )
 
     # build the plot
     pb1 <- ggplot2::ggplot_build(p1)
 
     # subtitle
     set.seed(123)
-    p1_subtitle <- statsExpressions::expr_t_bayes(
-      data = data_bugs_2,
-      x = condition,
-      y = desire,
-      k = 4,
-      paired = TRUE,
-      conf.level = 0.99,
-      messages = FALSE
-    )
+    p1_subtitle <-
+      statsExpressions::expr_t_bayes(
+        data = data_bugs_2,
+        x = condition,
+        y = desire,
+        k = 4,
+        paired = TRUE,
+        conf.level = 0.99,
+        messages = FALSE
+      )
 
     # dataframe used for visualization
     testthat::expect_equal(length(pb1$data), 8L)
@@ -296,7 +296,6 @@ testthat::test_that(
         pairwise.annotation = "p",
         outlier.tagging = FALSE,
         pairwise.comparisons = TRUE,
-        axes.range.restrict = TRUE,
         conf.level = 0.90,
         messages = FALSE
       )
@@ -455,15 +454,15 @@ testthat::test_that(
     library(WRS2)
 
     # plot
-    p <- ggstatsplot::ggwithinstats(
-      data = WineTasting,
-      x = Wine,
-      y = Taste,
-      axes.range.restrict = TRUE,
-      results.subtitle = FALSE,
-      messages = FALSE,
-      ggplot.component = ggplot2::labs(y = "Taste rating")
-    )
+    p <-
+      ggstatsplot::ggwithinstats(
+        data = WineTasting,
+        x = Wine,
+        y = Taste,
+        results.subtitle = FALSE,
+        messages = FALSE,
+        ggplot.component = ggplot2::labs(y = "Taste rating")
+      )
 
     # build plot
     pb <- ggplot2::ggplot_build(p)
@@ -538,16 +537,17 @@ testthat::test_that(
 
     # capture the message
     set.seed(123)
-    p <- suppressWarnings(ggstatsplot::ggwithinstats(
-      data = df,
-      x = x,
-      y = y,
-      pairwise.display = "significant",
-      pairwise.annotation = "p.value",
-      pairwise.comparisons = TRUE,
-      sphericity.correction = TRUE,
-      messages = FALSE
-    ))
+    p <-
+      suppressWarnings(ggstatsplot::ggwithinstats(
+        data = df,
+        x = x,
+        y = y,
+        pairwise.display = "significant",
+        pairwise.annotation = "p.value",
+        pairwise.comparisons = TRUE,
+        sphericity.correction = TRUE,
+        messages = FALSE
+      ))
 
     # build the plot
     pb <- ggplot2::ggplot_build(p)