+++ tests/testthat/test-grouped_ggbarstats.R
@@ -112,44 +112,77 @@ testthat::test_that(
         data = dplyr::sample_frac(tbl = forcats::gss_cat, size = 0.1),
         main = relig,
         condition = marital,
-        grouping.var = race,
+        grouping.var = "race",
         output = "subtitle",
-        results.subtitle = FALSE,
-        bar.proptest = FALSE,
-        messages = FALSE
-      ))
-
-    # tests
-    testthat::expect_equal(length(ls_results), 3L)
-    testthat::expect_null(ls_results[[1]], NULL)
-    testthat::expect_null(ls_results[[2]], NULL)
-    testthat::expect_null(ls_results[[3]], NULL)
-
-    # checking results
-    set.seed(123)
-    results_ls <-
-      suppressWarnings(ggstatsplot::grouped_ggbarstats(
-        data = as.data.frame(HairEyeColor),
-        main = Hair,
-        condition = Eye,
-        counts = "Freq",
-        bias.correct = FALSE,
-        grouping.var = Sex,
-        output = "subtitle",
+        k = 3,
         messages = FALSE
       ))
 
     # checking subtitle
-    testthat::expect_identical(
-      results_ls$Male,
-      ggplot2::expr(
-        paste(
+    testthat::expect_equal(
+      ls_results,
+      list(
+        Other = ggplot2::expr(paste(
           NULL,
           chi["Pearson"]^2,
           "(",
-          "9",
+          "40",
           ") = ",
-          "41.28",
+          "40.274",
+          ", ",
+          italic("p"),
+          " = ",
+          "0.458",
+          ", ",
+          widehat(italic("V"))["Cramer"],
+          " = ",
+          "0.009",
+          ", CI"["95%"],
+          " [",
+          "-0.295",
+          ", ",
+          "-0.020",
+          "]",
+          ", ",
+          italic("n")["obs"],
+          " = ",
+          182L
+        )),
+        Black = ggplot2::expr(paste(
+          NULL,
+          chi["Pearson"]^
+            2,
+          "(",
+          "32",
+          ") = ",
+          "25.113",
+          ", ",
+          italic("p"),
+          " = ",
+          "0.801",
+          ", ",
+          widehat(italic("V"))["Cramer"],
+          " = ",
+          "0.000",
+          ", CI"["95%"],
+          " [",
+          "-0.168",
+          ", ",
+          "0.006",
+          "]",
+          ", ",
+          italic("n")["obs"],
+          " = ",
+          317L
+        )),
+        White = ggplot2::expr(paste(
+          NULL,
+          chi["Pearson"]^
+            2,
+          "(",
+          "52",
+          ") = ",
+          "109.652",
           ", ",
           italic("p"),
           " = ",
@@ -157,112 +190,19 @@ testthat::test_that(
           ", ",
           widehat(italic("V"))["Cramer"],
           " = ",
-          "0.22",
+          "0.094",
           ", CI"["95%"],
           " [",
-          "0.14",
+          "0.032",
           ", ",
-          "0.26",
+          "0.100",
           "]",
           ", ",
           italic("n")["obs"],
           " = ",
-          279L
-        )
+          1649L
+        ))
       )
     )
-
-    testthat::expect_identical(
-      results_ls$Female,
-      ggplot2::expr(
-        paste(
-          NULL,
-          chi["Pearson"]^2,
-          "(",
-          "9",
-          ") = ",
-          "106.66",
-          ", ",
-          italic("p"),
-          " = ",
-          "< 0.001",
-          ", ",
-          widehat(italic("V"))["Cramer"],
-          " = ",
-          "0.34",
-          ", CI"["95%"],
-          " [",
-          "0.28",
-          ", ",
-          "0.38",
-          "]",
-          ", ",
-          italic("n")["obs"],
-          " = ",
-          313L
-        )
-      )
-    )
-  }
-)
-
-# checking if results coincide with base version -----------------------------
-
-testthat::test_that(
-  desc = "checking if results coincide with base version",
-  code = {
-    testthat::skip_on_cran()
-
-    # creating new datasets from the existing one
-    mtcars2 <- dplyr::mutate(mtcars, grp = "1")
-    mtcars3 <- dplyr::filter(mtcars2, cyl != "8")
-
-    set.seed(123)
-    p1 <-
-      suppressWarnings(
-        ggbarstats(
-          data = mtcars,
-          main = "am",
-          condition = cyl,
-          messages = FALSE,
-          output = "subtitle"
-        )
-      )
-
-    set.seed(123)
-    p2 <-
-      suppressWarnings(ggbarstats(
-        data = mtcars3,
-        x = "am",
-        y = cyl,
-        messages = FALSE,
-        output = "subtitle"
-      ))
-
-    set.seed(123)
-    p3 <-
-      suppressWarnings(grouped_ggbarstats(
-        data = mtcars2,
-        main = am,
-        y = "cyl",
-        grouping.var = grp,
-        messages = FALSE,
-        output = "subtitle"
-      ))
-
-    set.seed(123)
-    p4 <-
-      suppressWarnings(grouped_ggbarstats(
-        data = mtcars3,
-        x = "am",
-        condition = cyl,
-        grouping.var = "grp",
-        messages = FALSE,
-        output = "subtitle"
-      ))
-
-    # testing if grouped and base versions results are same
-    testthat::expect_identical(p1, p3$`1`)
-    testthat::expect_identical(p2, p4$`1`)
   }
 )